+++ src/main/java/net/replaceitem/discarpet/config/Config.java
@@ -7,9 +7,12 @@ import java.util.List;
 public class Config {
     @SerializedName("bots")
     public List<BotConfig> BOTS;
+    @SerializedName("connect_timeout")
+    public Integer CONNECT_TIMEOUT;
     
     public void fillDefaults() {
         if(BOTS == null) BOTS = List.of(new BotConfig());
         BOTS.forEach(BotConfig::fillDefaults);
+        if(CONNECT_TIMEOUT == null) CONNECT_TIMEOUT = 10;
     }
 }